Update and simplify samples + update dependency list
