+++ samples/fcm-sample/proguard-rules.pro
