+++ samples/fcm-sample/src/main/AndroidManifest.xml
