+++ samples/fcm-sample/src/main/res/drawable-xxhdpi/ic_chat.png
