+++ samples/fcm-sample/src/main/res/drawable-xxhdpi/ic_launcher.png
