+++ samples/fcm-sample/src/main/res/drawable/rounded_badge.xml
