+++ samples/fcm-sample/src/main/res/layout/activity_main.xml
