+++ samples/fcm-sample/src/main/res/values-w820dp/dimens.xml
