+++ samples/fcm-sample/src/main/res/values/color.xml
