+++ samples/fcm-sample/src/main/res/values/dimens.xml
