+++ samples/fcm-sample/src/main/res/values/styles.xml
