+++ samples/gcm-sample/proguard-rules.pro
