+++ samples/gcm-sample/src/main/AndroidManifest.xml
