+++ samples/gcm-sample/src/main/res/drawable-xxhdpi/ic_launcher.png
