+++ samples/gcm-sample/src/main/res/drawable/rounded_badge.xml
