+++ samples/gcm-sample/src/main/res/layout/activity_main.xml
