+++ samples/gcm-sample/src/main/res/values-w820dp/dimens.xml
