+++ samples/gcm-sample/src/main/res/values/color.xml
