+++ samples/gcm-sample/src/main/res/values/dimens.xml
