+++ samples/gcm-sample/src/main/res/values/strings.xml
