+++ samples/gcm-sample/src/main/res/values/styles.xml
