+++ samples/gradle.properties
