--- samples/intercom-sample-fcm/.gitignore
@@ -1,6 +1,0 @@
-.gradle
-local.properties
-.idea
-.DS_Store
-build
-*.iml
