--- samples/intercom-sample-fcm/app/build.gradle
@@ -1,28 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.0"
-
-    defaultConfig {
-        applicationId "io.intercom.sample"
-        minSdkVersion 15
-        targetSdkVersion 26
-        versionCode 1
-        versionName "1.0"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-dependencies {
-    compile 'io.intercom.android:intercom-sdk-base:4.+'
-    compile 'io.intercom.android:intercom-sdk-fcm:4.+'
-    compile 'com.google.firebase:firebase-messaging:11.0.4'
-}
-
-apply plugin: 'com.google.gms.google-services'
