--- samples/intercom-sample-fcm/settings.gradle
@@ -1,1 +1,0 @@
-include ':app'
