--- samples/intercom-sample/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Fri Jun 09 10:10:28 IST 2017
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-3.4.1-all.zip
