+++ samples/settings.gradle
@@ -1,0 +1,2 @@
+include ':fcm-sample'
+include ':gcm-sample'
